updated seveirty names
