+++ apps/server/src/services/logger/index.ts
@@ -1,1 +1,2 @@
+export * from './levels';
 export * from './logger';
