+++ apps/server/src/services/logger/levels.ts
@@ -1,0 +1,13 @@
+export enum LogLevels {
+  Error = 'error',
+  Warn = 'warn',
+  Info = 'info',
+  Debug = 'debug',
+}
+
+export const logLevels = {
+  [LogLevels.Error]: 0,
+  [LogLevels.Warn]: 1,
+  [LogLevels.Info]: 2,
+  [LogLevels.Debug]: 3,
+};
